More cleanup, fix typos in comments

+++ src/java.base/windows/classes/sun/nio/fs/WindowsFileAttributes.java
@@ -329,7 +329,7 @@ class WindowsFileAttributes
                 int size = MAXIMUM_REPARSE_DATA_BUFFER_SIZE;
                 try (NativeBuffer reparseBuffer = NativeBuffers.getNativeBuffer(size)) {
                     DeviceIoControl(handle, FSCTL_GET_REPARSE_POINT, reparseBuffer.address(), size);
-                    reparseTag = (int)unsafe.getLong(reparseBuffer.address());
+                    reparseTag = unsafe.getInt(reparseBuffer.address());
                 }
             }
